requirements.txt
flask
google-cloud-bigquery
google-cloud-bigquery-storage
pandas
plotly
beautifulsoup4
vertexai
db-dtypes